fix: show errors modal on adding bulk items to cart
